% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_reshape.R
\name{brick_to_matrix}
\alias{brick_to_matrix}
\title{brick_to_matrix}
\usage{
brick_to_matrix(b)
}
\arguments{
\item{b}{A raster \code{\link[raster]{brick}}.}
}
\value{
An object of class \code{\link[rsar]{SAR_matrix}}; a specialisation
of class matrix.
}
\description{
Convert a raster \code{\link[raster]{brick}} of
dimensions \code{A x B x C} into a matrix of
dimensions \code{A*B x C} while keeping
track of the geospatial and dimension attributes of the
original brick.
}
\examples{
filename <- system.file(
  "extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
b <- raster::brick(filename)
m <- brick_to_matrix(b)

}
