% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{choose_order}
\alias{choose_order}
\title{choose_order}
\usage{
choose_order(m, sample_size = 1000L)
}
\arguments{
\item{m}{A matrix.}

\item{sample_size}{An integer smaller than
\code{nrow(m)}.}
}
\value{
An integer.
B
}
\description{
Take a random sample (of size \code{sample_size})
of the rows of \code{m}, then fit an AR model to
each row in the sample, using
AIC to determine the optimal order.
Return the largest order across all sampled
rows.
}
\examples{
filepath <- system.file("extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
m <- load_SAR_matrix(filepath)
order <- choose_order(m)


}
