% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{fit_AR_to_SAR}
\alias{fit_AR_to_SAR}
\title{fit_AR_to_SAR}
\usage{
fit_AR_to_SAR(m, order = choose_order(m))
}
\arguments{
\item{m}{A matrix.}
}
\value{
A matrix of AR coefficients where each row corresponds to
one row of \code{m}.
}
\description{
Fit an AR model with the chosen \code{order}
to each row of \code{m} using \code{\link[base]{ar}}
(so each row of \code{m} is treated as a single
time series). The results are returned in a matrix whose
columns are the resulting AR coefficients.
}
\examples{
filepath <- system.file("extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
m <- load_SAR_matrix(filepath)
ar_m <- fit_AR_to_SAR(m)

}
