% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tif.R
\name{give_attrs}
\alias{give_attrs}
\title{give_attrs}
\usage{
give_attrs(br_b, br, incl_names = TRUE)
}
\arguments{
\item{br_b}{A raster brick}

\item{br}{A raster brick}

\item{incl_names}{Logical}
}
\description{
give_attrs
}
