% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is_SAR_matrix}
\alias{is_SAR_matrix}
\title{is_SAR_matrix}
\usage{
is_SAR_matrix(m)
}
\arguments{
\item{m}{An object}
}
\value{
Logical \code{TRUE} if \code{m} is a \code{\link[rsar]{SAR_matrix}}
object, otherwise \code{FALSE}.
}
\description{
Check if an object is of class \code{\link[rsar]{SAR_matrix}}.
}
\examples{
filename <- system.file(
  "extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
m <- load_SAR_matrix(filename)
is_SAR_matrix(m)

}
