% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tif.R
\name{matrix_to_brick}
\alias{matrix_to_brick}
\title{matrix_to_brick}
\usage{
matrix_to_brick(m)
}
\arguments{
\item{m}{An \code{SAR_matrix}  object, such as that output
by \code{\link[rsar]{load_SAR_matrix}}.}
}
\value{
A raster \code{\link[raster]{brick}}.
}
\description{
Convert an \code{\link[rsar]{SAR_matrix}} object \code{m},
such as that output by \code{\link[rsar]{load_SAR_matrix}}, into a
raster \code{\link[raster]{brick}} with the
geospatial and dimension attributes stored in \code{m}.
}
\examples{
filename <- system.file(
  "extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
m <- load_SAR_matrix(filename)
matrix_to_brick(m)

}
