% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_reshape.R
\name{patchify_SAR_matrix}
\alias{patchify_SAR_matrix}
\title{patchify_SAR_matrix}
\usage{
patchify_SAR_matrix(m, d, fun = mean)
}
\arguments{
\item{m}{An \code{SAR_matrix}  object, such as that output
by \code{\link[rsar]{load_SAR_matrix}}.}

\item{d}{The dimension of patches.}

\item{fun}{The aggregation function. Default is mean.}
}
\value{
A \code{\link[rsar]{SAR_matrix}} of reduced dimension,
with appropriately adjusted raster attributes.
}
\description{
Reduce the resolution of images (columns) in a
\code{\link[rsar]{SAR_matrix}} object \code{m},
by aggregating all values inside \eqn{d \times d} patches
using the aggregation function \code{fun}.
}
\examples{
filename <- system.file(
  "extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
m <- load_SAR_matrix(filename)
dim(m)
mp <- patchify_SAR_matrix(m,4)
head(mp)
dim(mp)
bp <- matrix_to_brick(mp)
raster::plot(bp)

}
