% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsar.R
\docType{package}
\name{rsar}
\alias{rsar}
\alias{rsar-package}
\title{Package title}
\description{
Description
}
\details{
Details
}
\section{Update}{
 Current updates are available through
URL: https://github.com/frycast/rsar
}

\section{BugReports}{
 https://github.com/frycast/rsar/issues
}

\author{
Daniel Fryer \email{daniel-fryer@live.com.au},
Mason Terrett \email{Mason.Terrett@gmail.com}
}
