% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_reshape.R
\name{brick_to_matrix}
\alias{brick_to_matrix}
\title{brick_to_matrix}
\usage{
brick_to_matrix(b, drop_na = TRUE, na_layer = 1L)
}
\arguments{
\item{b}{A raster \code{\link[raster]{brick}}.}

\item{drop_na}{A logical. If \code{TRUE} then any
\code{NA} pixels are dropped, but a record is kept of their
location so that the original raster brick can be easily
reconstructed later with \code{\link[rsar]{matrix_to_brick}}.
Note, to save time, only one layer is checked for \code{NA}
pixels, i.e., it is assumed that an \code{NA} pixel is
\code{NA} across all layers.}

\item{na_layer}{An integer specifying the index of the
layer to check for \code{NA} values if \code{drop_na}
is \code{TRUE}.}
}
\value{
An object of class \code{\link[rsar]{SAR_matrix}}; a specialisation
of class matrix.
}
\description{
Convert a raster \code{\link[raster]{brick}} of
dimensions \code{A x B x C} into a matrix of
dimensions \code{A*B x C} while keeping
track of the geospatial and dimension attributes of the
original brick.
}
\examples{
filename <- system.file(
  "extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
b <- raster::brick(filename)
m <- brick_to_matrix(b)

}
