% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_reshape.R
\name{load_SAR_matrix}
\alias{load_SAR_matrix}
\title{load_SAR_matrix}
\usage{
load_SAR_matrix(filename, drop_na = TRUE)
}
\arguments{
\item{filename}{This is usually a string giving absolute or relative
path to a \code{tif} file containing a SAR raster brick. This argument is
passed to \code{\link[raster]{brick}}.}

\item{drop_na}{A logical. If \code{TRUE} then any
\code{NA} pixels are dropped, but a record is kept of their
location so that the original raster brick can be easily
reconstructed later with \code{\link[rsar]{matrix_to_brick}}.
Note, to save time, only one layer is checked for \code{NA}
pixels, i.e., it is assumed that an \code{NA} pixel is
\code{NA} across all layers.}
}
\value{
An object of class \code{\link[rsar]{SAR_matrix}}; a specialisation
of class matrix.
}
\description{
Given a file path to a SAR raster image
in \code{tif} format, convert the raster brick of
dimensions \code{A x B x C} into a matrix of
dimensions \code{A*B x C} while keeping
track of the geospatial and dimension attributes of the
original \code{\link[raster]{brick}}.
}
\examples{
filename <- system.file(
  "extdata", "MG_CC_sub_norm_testclip.tif", package="rsar")
load_SAR_matrix(filename)

}
